apps/mcp-drive/pyproject.toml
[project]
name = "mcp-drive"
version = "0.1.0"
description = "MCP server: awareness READ-ONLY do Google Drive Thórus pra agentes do Oráculo."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Thórus Engenharia" },
]
dependencies = [
    "mcp>=1.2.0,<2.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "google-auth>=2.30.0,<3.0.0",
    "google-api-python-client>=2.140.0,<3.0.0",
]

[project.scripts]
mcp-drive = "mcp_drive.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_drive"]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "ASYNC"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src", "tests"]
